Extract provider model controls
