+++ tests/test-custom-api.js
@@ -217,7 +217,8 @@ console.log('\n12. settings.js + provider panel source inspection');
 const settingsSrc = read('js/settings.js');
 const panelsSrc = read('js/provider-panels.js');
 const panelRenderSrc = read('js/provider-panel-renderers.js');
-const providerUiSrc = panelsSrc + panelRenderSrc;
+const providerModelControlsSrc = read('js/provider-model-controls.js');
+const providerUiSrc = panelsSrc + panelRenderSrc + providerModelControlsSrc;
 assert('settings.js has data-provider="custom" button', settingsSrc.includes('data-provider="custom"'));
 assert("settings.js wires switchAIProvider('custom')", settingsSrc.includes("switchAIProvider('custom')"));
 assert('provider code imports getCustomApiUrl', providerUiSrc.includes('getCustomApiUrl'));
@@ -225,14 +226,14 @@ assert('provider-panels imports setCustomApiUrl', panelsSrc.includes('setCustomA
 assert('provider code imports getCustomApiKey', providerUiSrc.includes('getCustomApiKey'));
 assert('provider-panels imports saveCustomApiKey', panelsSrc.includes('saveCustomApiKey'));
 assert('provider code imports getCustomApiModel', providerUiSrc.includes('getCustomApiModel'));
-assert('provider-panels imports setCustomApiModel', panelsSrc.includes('setCustomApiModel'));
+assert('provider-model-controls imports setCustomApiModel', providerModelControlsSrc.includes('setCustomApiModel'));
 assert('provider-panels imports fetchCustomApiModels', panelsSrc.includes('fetchCustomApiModels'));
 assert('provider-panels imports validateCustomApiKey', panelsSrc.includes('validateCustomApiKey'));
 assert('renderAIProviderPanel handles custom', panelRenderSrc.includes("provider === 'custom'"));
 assert('handleSaveCustomApi exists', panelsSrc.includes('function handleSaveCustomApi()'));
 assert('handleRemoveCustomApi exists', panelsSrc.includes('function handleRemoveCustomApi()'));
-assert('renderCustomApiModelDropdown exists', panelsSrc.includes('function renderCustomApiModelDropdown('));
-assert('applyCustomApiManualModel exists', panelsSrc.includes('function applyCustomApiManualModel()'));
+assert('renderCustomApiModelDropdown exists', providerModelControlsSrc.includes('function renderCustomApiModelDropdown('));
+assert('applyCustomApiManualModel exists', providerModelControlsSrc.includes('function applyCustomApiManualModel()'));
 assert('custom-url-input element', panelRenderSrc.includes('custom-url-input'));
 assert('custom-key-input element', panelRenderSrc.includes('custom-key-input'));
 assert('custom-model-area element', providerUiSrc.includes('custom-model-area'));
